+++ crates/testing/test-programs/js/webhook/call_activity.js
@@ -1,5 +1,7 @@
 
-// Webhook that calls an activity using obelisk.call
+import * as dynamic from "obelisk:webhook-dynamic@1.0.0";
+
+// Webhook that calls an activity dynamically
 export default function handle(request) {
     const a = Number(process.env['a']);
     const b = Number(process.env['b']);
@@ -12,6 +14,6 @@ export default function handle(request) {
 
 function call(a, b) {
     return function (a, b) {
-        return obelisk.call("testing:integration/activity.add", [a, b]);
+        return dynamic.call("testing:integration/activity.add", [a, b]);
     }(a, b)
 }